[e.cash] Fix non-clickable button on /get-ecash
Summary:
Bottom of /get-ecash in the staking section the button was not clickable in some cases due to the
absolutely positioned image background overlapping it. Adding a higher z-index to bring it up and make
clickable

Test Plan: preview the site and got the staking section of /get-ecash. The button should work and go to the staking page

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16041