[Cashtab Extension] Increase height to improve app->extension UX
Summary:
The recent additions of the cashtab message and opreturnraw toggles plus the output textarea have pushed the Send button to just beyond the existing dimensions, causing the users to scroll down to Send.

This diff increases the height to facilitate a smoother app->extension UX.

Current NOTIFICATION_HEIGHT @ 750
{F7535508}

New NOTIFICATION_HEIGHT @ 950
{F7535513}

Test Plan: Use an external app to trigger a BIP21 transaction to the extension and observe the Send button is in view without needing to scroll.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16047