Reduce port range in functional tests
Summary:
So that we never use port number > 32000.
Docker uses ports 32768-65535 for its own use (containerd) so there is a risk of collision that can be avoided by reducing the range.

Test Plan:
In various conditions, with chronik enabled (the only tests that can reach ports > 28000):
  ninja check-functional

See CI calls in the comments in this diff.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16048