[ecash-lib] Add ALP support
Summary:
Add functions for building GENESIS, MINT, SEND, BURN section pushdata, and to build eMPP OP_RETURN Scripts.

It doesn't include any wallet functionality, this is for users of this library to implement (or for a built-in wallet in a future diff).

`GenesisInfo` is based on the `GenesisInfo` of chronik-client.

Test Plan: `npm run integration-tests`

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16049