[Cashtab] Upgrade NFT Page and add support for sending NFTs
Summary:
Upgrade token cache to get `groupTokenId` for NFTs. This info is available in the `tx` call and never changes, so it should be cached.

Modify the `token/` page so that, if a user visits without `groupTokenId` in cache, Cashtab will fetch this info and update the cache with it. Also ensure that Cashtab will fetch cached token info for the nft parent token (now that we support sends, it is possible for a Cashtab user to receive an NFT child while not having the parent in cache).

The cache changes could have stood alone, but I started this diff to add the NFT send feature, and discovered these would be needed first in the process of adding. Since the functions for sending already exist, and we are already upgrading the NFT UI, go ahead and implement Send here as well.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16060