move DummyConfig to test/util/setup_common.h
Summary:
This is a test only class, so it makes sense to have it defined in the test directory.

In a following commit, I want to add an `ArgsManager` argument for `CreateChainParams`, and this will avoid having to couple config.h/cpp with the ArgsManager (system.h). This is needed for decoupling CChainParams and ArgsManager in future Core backports.

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16108