[doc] Minor improvments on the teamcity agent readme
Summary:
Debian Stretch is too old for being recommended anymore, and Ubuntu is untested so remove them for simplicity.
Bump the memory requirements for running the TSAN builds.
The disk size is left untouched as it's enough for some builds, but clearly won't cut it for most, and as mentioned it can be extended.

Depends on D16127.

Test Plan: Read it.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16128