[Cashtab] Stop using Title for Tx History
Summary:
Earlier, I went through Cashtab and got rid of all `data-testid`. However, there are still some cases where it makes sense.

For example here -- we want to make sure the app is rendering Tx History and not the expected content when a wallet has no history. React testing library would best practices prefer us to make sure the "right" tx is rendered, we could check for example by confirming the Amount Sent is as expected for the first tx. But testing to this level of specificity is unduly complicated -- we only want to confirm that tx history is rendered (if we change the mock wallet or mock tx history, we do not want to have to update all the tests).

So, just use data-testid. This way we do not render an annoying tooltip when mouse hovers over any tx.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16130