util: move util/message to common/signmessage
Summary:
Move util/message to common/signmessage so it is named more clearly, and
because the util library is not supposed to depend on other libraries besides
the crypto library. The signmessage functions use CKey, CPubKey, PKHash, and
DecodeDestination functions in the consensus and common libraries.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/29015 | core#29015]]
https://github.com/bitcoin/bitcoin/pull/29015/changes/6861f954f8ff42c87ad638037adae86a5bd89600

Depends on D19742

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19743