[Cashtab] Show exact XEC amount sent or received in XEC tx history if user clicks tx
Summary:
T3570

When glancing of txs, the exact amount is often less useful than the formatted / rounded amount. However, it is annoying that user cannot see the exact amount.

Show the exact amount if the user clicks the tx, which is existing behavior for showing more info/options about this tx.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16160