indexes: Don't wipe indexes again when already reindexing
Summary:
Before this change continuing a reindex without the -reindex flag set
would leave the block and coins db intact, but discard the data of the
optional indexes. While not a bug per se, wiping the data again is
wasteful, both in terms of having to write it again, and potentially
leading to longer startup times.

When initially running a reindex, both the block index and any further
activated indexes are wiped. On an index's Init(), both the best block
stored by the index and the chain's tip are null. An index's m_synced
member is therefore true. This means that it will process blocks through
validation events while the reindex is running.

Currently, if the reindex is continued without the user re-specifying
the reindex flag, the block index is preserved but further index data is
wiped. This leads to the stored best block being null, but the chain tip
existing. The m_synced member will be set to false. The index will not
process blocks through the validation interface, but instead use the
background sync once the reindex is completed.

If the index is preserved (this change) after a restart its best block
may potentially match the chain tip. The m_synced member will be set to
true and the index can process validation events during the rest of the
reindex.

Co-authored-by: furszy <matiasfurszyfer@protonmail.com>
This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/30132 | core#30132]]
https://github.com/bitcoin/bitcoin/pull/30132/changes/201c1a92824c71ae646d5bba9963871b1d704cc1
https://github.com/bitcoin/bitcoin/pull/30132/changes/1b1c6dcca0cc891bd35d29b61628c39098cd94ce
Depends on D20602

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20603