refactor: Move system from util to common library
Summary:
Since the kernel library no longer depends on the system file, move it
to the common library instead in accordance to the diagram in
doc/design/libraries.md.

This is a move-only commit.

This concludes backport of [[https://github.com/bitcoin/bitcoin/pull/27636 | core#27636]]
https://github.com/bitcoin/bitcoin/pull/27636/commits/7d3b35004b039f2bd606bb46a540de7babdbc41e
Depends on D16216

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16217