[Cashtab] Show full-height screen with no nav footer for onboarding page
Summary:
Hangover from earlier CSS scroll head patches in Cashtab. The height of hte wallet body component is set to leave exact space for the footer, so that we have no unnecessary scrolling.

On the onboarding page, we disable the footer, as we do not want navigation to any screens until the user has a wallet.

Adjust height of container for this case so there is not an empty space where the footer should be.

Test Plan:
`npm start`, open `localhost:3000` in an incognito window, note there is no gap at the bottom, create a wallet, note footer is present

before:
{F7666679}

after:
{F7666681}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D16237