[avalanche] pass a reference to the avalanche processor to StakingRewardsPolicy
Summary:
This removes a few callsites for the global g_avalanche. It makes sense to assume that a StakingRewardPolicy should be used only when the processor is initialized, passing a reference instead of invoking a pointer makes that a hard rule.

Note that the is `IsStakingRewardsActivated` function is always called after we already checked for `g_avalanche` in non-test code, so there is no change in behavior by removing that check. And this allows removing some setup code in stakingrewards_tests.cpp

Depends on D16259

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16260