[Cashtab] Modify extension manifest to support wasm
Summary:
Chrome extensions do not automatically allow wasm execution. This must be specified in a CSP directive in the manifest.

Ref https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy

The `wasm-unsafe-eval` directive, while scary looking, "Allows the loading and execution of WebAssembly modules without the need to also allow unsafe JavaScript execution"

The web assembly is a bit of a black box -- however we are building it from a trusted source.

Test Plan: `npm run extension`, load manually, it works

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16261