[avalanche] pass the processor to Chainstate::ActivateBestChain[Step]
Summary:
Another step in removing access to the g_avalanche global in Chainstate code.

This adds `g_avalanche` access in rpc/blockchain.cpp, but this is only temporary because in a later diff the processor will be avaialble through  the node context.

Depends on D16265

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16266