test: Add ReceiveWithExtraTransactions Compact Block receive test.
Summary:
This new test uses the `vExtraTxnForCompact` (`extra_txn`) vector of
optional orphan/conflicted/etc. transactions to provide a transaction
in a compact block that was not otherwise present in our mempool.

This also covers an improbable nullptr deref bug addressed in #29752 where the
`extra_txn` vec/circular-buffer was sometimes null-initialized and
not yet filled when dereferenced in `PartiallyDownloadedBlock::InitData`.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/30237 | core#30237]]
https://github.com/bitcoin/bitcoin/pull/30237/commits/4c99301220ab44e98d0d0e1cc8d774d96a25b7aa

Depends on D19243

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19244