[explorer] Show the token balance
Summary:
This was not working because of a change in the token id format from the NNG chronik (previously was bytes, now it's hex encoded). This caused a double hex encoding issue that failed the token lookup.

Ref T3543.

Test Plan:
  cargo run
Check an address page with tokens, the balance is correctly displayed.

Reviewers: #bitcoin_abc, Mengerian

Reviewed By: #bitcoin_abc, Mengerian

Subscribers: Mengerian

Maniphest Tasks: T3543

Differential Revision: https://reviews.bitcoinabc.org/D16304