[avalanche] pass fAutoUnpark to Chainstate::FindMostWorkChain
Summary:
and fill blocksToReconcile unconditionnaly.

This removes the need for Chainstate::FindMostWorkChain to know anything about avalanche.

In a following diff, the  `isAvalancheEabled` function will be removed in favor of checking whether the avalanche processor pointer is not nullptr, so better to move the checks for avalanche in a scope that already has access to the pointer.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16311