Revert "[chronik] Disable the Write Ahead Log for rocksdb"
Summary:
This reverts commit rABC9abcbdc2ec3bc8762996ed58f94e7f6a41ee3fa5 and reverts commit rABCe98e1a1ccd5dab8607ee1232ec41b12285822a35.

This fixed the qemu failures at the cost of a lower db resiliency.
Now that the code evolved a bit this might not be required anymore and
we can avoid this tradeoff.

We also need to use the main qemu version as the backport version throwns bus errors on arm sometimes.

Test Plan: Spam the tests using qemu

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16322