[Cashtab] Remove "webapp tx request" warning
Summary:
Intention of this warning is that anyone could make a link that populates a tx into Cashtab.

We should show where the tx is coming from. However, this is non-trivial to do in a way that is itself immune from spoofing.

At the moment we are fighting a hypothetical problem. We do not know if the warning is effective. We do know that it is unsightly.

As with many crypto issues, with great power comes great responsibility. For Cashtab, smoother UX is always more important than security theater.

We do not have any specific examples of a user who was "saved" by this warning, which does not prevent the user from completing a tx anyway. So, get rid of it.

Test Plan:
`npm test`
`npm run extension`, use the local extension and note absence of warning

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D16324