[alias-server] Migrate to in-node chronik-client
Summary:
With ecash-chat, there is a use case for aliases. While it would be nice to launch aliases as NFTs for instant trading, the alias spec should still be leveraged for initial registrations.

Bring alias-server up to speed with the latest chronik-client as prep for a limited launch.

Note: the in-node version of chronik-client offers other ways to significantly improve this indexer. We could, for example, use a websocket subscription to listen for txs with the right LOKAD, then validate them on the server, finalize them on `TX_FINALIZED` msg, or use some kind of a plugin to index these txs. For now though, just getting the equivalent performance (without the weird wait and repeat RPC finalization check call) is step one.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16344